backend/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
celery
redis
python-dotenv
requests
django-cors-headers
social-auth-app-django